[IDP-2296] refactor UserInitiativeCountersUnlockMediatorService
