+++ src/main/java/it/gov/pagopa/reward/connector/event/consumer/TransactionConsumerConfig.java
@@ -1,6 +1,5 @@
 package it.gov.pagopa.reward.connector.event.consumer;
 
-import it.gov.pagopa.reward.service.counters.UserInitiativeCountersUnlockMediatorService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -13,7 +12,7 @@ import java.util.function.Consumer;
 @Slf4j
 public class TransactionConsumerConfig {
     @Bean
-    public Consumer<Flux<Message<String>>> trxResponseConsumer(UserInitiativeCountersUnlockMediatorService userInitiativeCountersUnlockMediatorService){
-        return userInitiativeCountersUnlockMediatorService::execute;
+    public Consumer<Flux<Message<String>>> trxResponseConsumer(TransactionConsumer transactionConsumer){
+        return transactionConsumer::execute;
     }
 }
