+++ src/test/java/it/gov/pagopa/reward/service/counters/UserInitiativeCountersUnlockMediatorServiceTest.java
@@ -1,12 +1,9 @@
 package it.gov.pagopa.reward.service.counters;
 
-import com.fasterxml.jackson.databind.ObjectReader;
 import com.mongodb.MongoException;
-import it.gov.pagopa.common.utils.TestUtils;
 import it.gov.pagopa.reward.connector.repository.UserInitiativeCountersRepository;
 import it.gov.pagopa.reward.dto.trx.RewardTransactionDTO;
 import it.gov.pagopa.reward.model.counters.UserInitiativeCounters;
-import it.gov.pagopa.reward.service.RewardErrorNotifierService;
 import it.gov.pagopa.reward.test.fakers.RewardTransactionDTOFaker;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -15,14 +12,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.messaging.Message;
-import org.springframework.messaging.support.MessageBuilder;
 import reactor.core.publisher.Mono;
-
-import java.time.Duration;
-import java.util.Collections;
-import java.util.Map;
-import java.util.function.Consumer;
 
 import static it.gov.pagopa.reward.utils.RewardConstants.PAYMENT_STATE_AUTHORIZED;
 import static it.gov.pagopa.reward.utils.RewardConstants.REWARD_STATE_REJECTED;
@@ -30,38 +20,27 @@ import static org.mockito.ArgumentMatchers.any;
 
 @ExtendWith(MockitoExtension.class)
 class UserInitiativeCountersUnlockMediatorServiceTest {
-    private static final long COMMIT_MILLIS = 10L;
-
     @Mock
     UserInitiativeCountersRepository userInitiativeCountersRepositoryMock;
-
-    @Mock
-    RewardErrorNotifierService rewardErrorNotifierServiceMock;
 
     UserInitiativeCountersUnlockMediatorServiceImpl userInitiativeCountersUnlockMediatorServiceImpl;
 
     @BeforeEach
     void setUp() {
         userInitiativeCountersUnlockMediatorServiceImpl = new UserInitiativeCountersUnlockMediatorServiceImpl(
-                "appName",
-                userInitiativeCountersRepositoryMock,
-                rewardErrorNotifierServiceMock,
-                COMMIT_MILLIS,
-                TestUtils.objectMapper);
+                userInitiativeCountersRepositoryMock);
     }
 
     @Test
     void execute_statusAcceptedAuthorized(){
-        RewardTransactionDTO payload = RewardTransactionDTOFaker.mockInstance(1);
-        payload.setStatus(PAYMENT_STATE_AUTHORIZED);
-        Message<String> message = MessageBuilder.withPayload("DUMMY PAYLOAD MESSAGE").build();
-        Map<String, Object> ctx = Collections.emptyMap();
+        RewardTransactionDTO trx = RewardTransactionDTOFaker.mockInstance(1);
+        trx.setStatus(PAYMENT_STATE_AUTHORIZED);
 
         UserInitiativeCounters userInitiativeCounters = new UserInitiativeCounters();
-        Mockito.when(userInitiativeCountersRepositoryMock.unlockPendingTrx(payload.getId()))
+        Mockito.when(userInitiativeCountersRepositoryMock.unlockPendingTrx(trx.getId()))
                         .thenReturn(Mono.just(userInitiativeCounters));
 
-        UserInitiativeCounters result = userInitiativeCountersUnlockMediatorServiceImpl.execute(payload, message, ctx).block();
+        UserInitiativeCounters result = userInitiativeCountersUnlockMediatorServiceImpl.execute(trx).block();
 
         Assertions.assertNotNull(result);
         Assertions.assertEquals(userInitiativeCounters, result);
@@ -71,13 +50,10 @@ class UserInitiativeCountersUnlockMediatorServiceTest {
 
     @Test
     void execute_statusNotAccepted(){
-        RewardTransactionDTO payload = RewardTransactionDTOFaker.mockInstance(1);
-        payload.setStatus(REWARD_STATE_REJECTED);
-        Message<String> message = MessageBuilder.withPayload("DUMMY PAYLOAD MESSAGE").build();
-        Map<String, Object> ctx = Collections.emptyMap();
+        RewardTransactionDTO trx = RewardTransactionDTOFaker.mockInstance(1);
+        trx.setStatus(REWARD_STATE_REJECTED);
 
-
-        UserInitiativeCounters result = userInitiativeCountersUnlockMediatorServiceImpl.execute(payload, message, ctx).block();
+        UserInitiativeCounters result = userInitiativeCountersUnlockMediatorServiceImpl.execute(trx).block();
 
         Assertions.assertNull(result);
         Mockito.verify(userInitiativeCountersRepositoryMock, Mockito.never()).unlockPendingTrx(any());
@@ -85,66 +61,20 @@ class UserInitiativeCountersUnlockMediatorServiceTest {
 
     @Test
     void execute_withException(){
-        RewardTransactionDTO payload = RewardTransactionDTOFaker.mockInstance(1);
-        payload.setStatus(PAYMENT_STATE_AUTHORIZED);
-        Message<String> message = MessageBuilder.withPayload("DUMMY PAYLOAD MESSAGE").build();
-        Map<String, Object> ctx = Collections.emptyMap();
+        RewardTransactionDTO trx = RewardTransactionDTOFaker.mockInstance(1);
+        trx.setStatus(PAYMENT_STATE_AUTHORIZED);
+
 
         String errorMessage = "DUMMY MONGO EXCEPTION";
-        Mockito.when(userInitiativeCountersRepositoryMock.unlockPendingTrx(payload.getId()))
+        Mockito.when(userInitiativeCountersRepositoryMock.unlockPendingTrx(trx.getId()))
                 .thenThrow(new MongoException(errorMessage));
 
-        Mono<UserInitiativeCounters> execute = userInitiativeCountersUnlockMediatorServiceImpl.execute(payload, message, ctx);
+        Mono<UserInitiativeCounters> execute = userInitiativeCountersUnlockMediatorServiceImpl.execute(trx);
         MongoException mongoException = Assertions.assertThrows(MongoException.class, execute::block);
 
         Assertions.assertEquals(errorMessage, mongoException.getMessage());
 
         Mockito.verifyNoMoreInteractions(userInitiativeCountersRepositoryMock);
-    }
-
-    @Test
-    void getCommitDelay(){
-        Duration result = userInitiativeCountersUnlockMediatorServiceImpl.getCommitDelay();
-
-        Assertions.assertEquals(Duration.ofMillis(COMMIT_MILLIS), result);
-    }
-
-    @Test
-    void getObjectReader(){
-        ObjectReader objectReaderResult = userInitiativeCountersUnlockMediatorServiceImpl.getObjectReader();
-
-        Assertions.assertNotNull(objectReaderResult);
-    }
-
-    @Test
-    void getFlowName(){
-        String flowNameResult = userInitiativeCountersUnlockMediatorServiceImpl.getFlowName();
-
-        Assertions.assertNotNull(flowNameResult);
-        Assertions.assertEquals("USER_COUNTER_UNLOCK", flowNameResult);
-    }
-
-    @Test
-    void onDeserializationError(){
-        Message<String> dummyMessage = MessageBuilder.withPayload("DUMMY_MESSAGE").build();
-
-        Consumer<Throwable> result = userInitiativeCountersUnlockMediatorServiceImpl.onDeserializationError(dummyMessage);
-
-        Assertions.assertNotNull(result);
-    }
-
-    @Test
-    void notifyError(){
-        Message<String> dummyMessage = MessageBuilder.withPayload("DUMMY_MESSAGE").build();
-        RuntimeException exception = new RuntimeException("DUMMY_EXCEPTION");
-
-
-        Mockito.when(rewardErrorNotifierServiceMock.notifyTransactionResponse(dummyMessage, "[USER_COUNTER_UNLOCK] An error occurred evaluating transaction", true, exception))
-                .thenReturn(true);
-
-        userInitiativeCountersUnlockMediatorServiceImpl.notifyError(dummyMessage, exception);
-
-        Mockito.verifyNoMoreInteractions(userInitiativeCountersRepositoryMock, rewardErrorNotifierServiceMock);
     }
 
 }
